[move] apiPayload패키지를 global 패키지로 이동 #3
