+++ src/main/java/chungbazi/chungbazi_be/global/apiPayload/code/BaseErrorCode.java
@@ -1,4 +1,4 @@
-package chungbazi.chungbazi_be.apiPayload.code;
+package chungbazi.chungbazi_be.global.apiPayload.code;
 
 public interface BaseErrorCode {
