+++ src/main/java/chungbazi/chungbazi_be/global/apiPayload/exception/ExceptionAdvice.java
@@ -1,8 +1,8 @@
-package chungbazi.chungbazi_be.apiPayload.exception;
+package chungbazi.chungbazi_be.global.apiPayload.exception;
 
-import chungbazi.chungbazi_be.apiPayload.ApiResponse;
-import chungbazi.chungbazi_be.apiPayload.code.ErrorReasonDTO;
-import chungbazi.chungbazi_be.apiPayload.code.status.ErrorStatus;
+import chungbazi.chungbazi_be.global.apiPayload.ApiResponse;
+import chungbazi.chungbazi_be.global.apiPayload.code.ErrorReasonDTO;
+import chungbazi.chungbazi_be.global.apiPayload.code.status.ErrorStatus;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.ConstraintViolationException;
 import lombok.extern.slf4j.Slf4j;
